Looking into the module method of coding
The buckygame code the second rewrite around is pretty nice, but I feel
like I can do much better, especially after having a lot of experience
now with a lot of javascript's inner workings, such as anonymous
functions and using the module style of coding. Looking to implement a
major rewrite of the engine ( again ) with layers for the map.